Update README.md
I was not able to get the code to work from this repository. However, I saved an archive that does build. 

Unfortunately, I don't think we can upload the working archive to Github because it contains some copyrighted code and there are also a number of binary files. I hope the updates to this readme will encourage someone who wants to run this client in the future to contact me, rather than trying to get the code in here running.